x86/vmx: Allow bit 13 of VM-entry injection info to be set
With the introduction of VMX nested-exception support, bit 13 of the
VM-entry interruption-information field (Nested Exception) must be set
when injecting a nested hardware exception.

A CPU supporting FRED relaxes the Intel SDM's conventional VM-entry
validation checks: if the event type is a hardware exception (bits 10:8
set to 3), bit 13 is permitted to be 1, regardless of whether FRED
transitions are actively enabled following the VM-entry.

Update the event injection test logic to allow this bit to be set to
prevent spurious test failures on FRED-capable hardware.

Signed-off-by: Xin Li <xin@zytor.com>