x86: eventinj: Allow Bit 13 of VM-entry intr info to be set
With the introduction of VMX nested-exception support, bit 13 of the
VM-entry intr info field needs to be set to 1 when a nested hardware
exception is injected.

Regardless of whether FRED transitions would be enabled following
VM-entry, a CPU supporting FRED relaxes the following check: if event
type (bits 10:8) is 3 (hardware exception), bit 13 may have value 1
(indicating a nested exception).

Signed-off-by: Xin Li (Intel) <xin@zytor.com>